response to dna.Base comments
